Rename main module to core
