+++ topothehourbot/__init__.py
@@ -1,1 +1,1 @@
-from .main import *
+from .core import *
